requirements.txt
numpy
coverage
pytest-coverage
tox